pyproject.toml
[project]
name = "ai-office-pilot"
version = "3.0.0"
description = "AI Office Assistant - Email, Files, Data automation with self-learning"
requires-python = ">=3.8"
dependencies = [
    "ollama>=0.4.0",
    "requests>=2.31.0",
    "imapclient>=3.0.0",
    "pyzmail36>=1.0.4",
    "python-docx>=1.1.0",
    "PyPDF2>=3.0.0",
    "Pillow>=10.0.0",
    "python-magic>=0.4.27",
    "watchdog>=4.0.0",
    "openpyxl>=3.1.0",
    "gspread>=6.0.0",
    "oauth2client>=4.1.3",
    "cryptography>=42.0.0",
    "argon2-cffi>=23.1.0",
    "flask>=3.0.0",
    "rich>=13.7.0",
    "click>=8.1.0",
    "plyer>=2.1.0",
    "python-dotenv>=1.0.0",
    "schedule>=1.2.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]
docker = [
    "docker>=6.0.0",
    "docker-compose>=1.29.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py38"
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "C",     # flake8-comprehensions
    "B",     # flake8-bugbear
    "UP",    # pyupgrade
    "SIM",   # flake8-simplify
    "RUF",   # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
]

[tool.ruff.lint.isort]
known-first-party = ["core", "security", "learning", "modules", "notifications"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["Tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["."]
omit = [
    "Tests/*",
    "venv/*",
    ".venv/*",
    "*/__pycache__/*",
    "data/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
